Communication rework.
Changed the controller script to not initialize a class or track state for reliable operation. Classes objects take a lot of space, and the board was already constrained. State seemd to have some issues tracking and eventually crashing. Instead letting the roomba itself handle the states and if commands are valid board is just passing the commands it picks up from the radio. Added some led signalling for external observation. Updated makefile to transfer code.py